Copying Files
This commit copies files over from the book recommendation app. Again, this repo is just creating a gh-pages for the book recommendation app. I am doing this seperately because I want the book recommendation app to have both a gh-pages and a dedicated domain that it points two (hence, two different repos).